requirements.txt
Django>=5.1,<5.2
wagtail>=6.4,<6.5

django-modelcluster~=6.4